set var

+++ buildnativeoperations.bat
@@ -1,12 +1,14 @@
 SET OMP_NUM_THREADS=1
+set argC=0
+for %%x in (%*) do set /A argC+=1
 
 
-if  "%#" LSS 1  (
-    echo "Please specify an argument"
+if  %argC% LSS 1  (
+    echo Please specify an argument
 ) else (
-    SET command="%1"
-    echo "Running %1"
-    if  "%1" == "clean"  (
+    SET command=%1
+    echo Running %1
+    if  %1 == clean  (
        rd /s /qcmake_install.cmake
        rd /s /qcubinbuild
        rd /s /qptxbuild
@@ -15,11 +17,11 @@ if  "%#" LSS 1  (
        rd /s /qtestbuild
        rd /s /qeclipse/CMake)les
 
-       echo "Deleted build"
+       echo Deleted build
     ) else  (
-         if "%1" ==  "eclipse"  (
+         if %1 ==  eclipse  (
             cd eclipse
-            cmake -G"Eclipse CDT4 - Unix Make)les" -DCMAKE_ECLIPSE_GENERATE_SOURCE_PROJECT=TRUE ..
+            cmake -GEclipse CDT4 - Unix Make)les -DCMAKE_ECLIPSE_GENERATE_SOURCE_PROJECT=TRUE ..
             python ./nsight-err-parse-patch.py ./project
             move eclipse/.cproject .
             move eclipse/.project .
@@ -27,7 +29,7 @@ if  "%#" LSS 1  (
          )
      else  (
 
-     if  "%1" ==  "lib"   (
+     if  %1 ==  lib   (
          rd /s /qlibrary  build
          mkdir librarybuild
          cd librarybuild
@@ -36,15 +38,15 @@ if  "%#" LSS 1  (
          )
         )
      else (
-         if "%1" ==  "test"   (
-           if  "%#" GTR "1"  
+         if %1 ==  test   (
+           if  %argC% GTR 1  
                 rd /s /qtestbuild
                 mkdir testbuild
                 cd testbuild
                 cmake -DRUN_TEST=TRUE ..
                 make && cd ..
                 move testbuild/test/libnd4jtests .
-               ./libnd4jtests -n "%2"
+               ./libnd4jtests -n %2
            else
                rd /s /qtestbuild
                mkdir testbuild
@@ -56,20 +58,20 @@ if  "%#" LSS 1  (
               )
            )
 
-           echo "FNISHING BUILD"
+           echo FNISHING BUILD
      else  (
-         if  "%1" == "cubin"  (
+         if  %1 == cubin  (
             rd /s /qcubinbuild
            mkdir cubinbuild
            cd cubinbuild
            cmake -DCUBIN=TRUE ..
            make && cd ..
-           echo "FNISHING BUILD"
+           echo FNISHING BUILD
            )
            )
            move cubinbuild/cubin/cuda_compile_cubin_generated_all.cu.cubin all.cubin
       else  (
-          if "%1" == "buffer"  (
+          if %1 == buffer  (
             rd /s /qbufferbuild
            mkdir bufferbuild
            cd bufferbuild
@@ -77,24 +79,24 @@ if  "%#" LSS 1  (
            make && cd ..
            )
            )
-           echo "FINISHING BUILD"
+           echo FINISHING BUILD
      else (
 
-         if "%1" == "blas"   (
+         if %1 == blas   (
             rd /s /qblasbuild
            mkdir blasbuild
            cd blasbuild
-           if  "%#" GTR "1"  
-              if  "%2" == "cuda"  
+           if  %# GTR 1  
+              if  %2 == cuda  
                    cmake -DCUDA_BLAS=true -DBLAS=TRUE ..
                    make && cd ..
-                  echo "FINISHING BUILD"
-              elif  "%2" == "cpu"  
+                  echo FINISHING BUILD
+              elif  %2 == cpu  
                    cmake -DCPU_BLAS=true -DBLAS=TRUE ..
                    make && cd ..
-                   echo "FINISHING BUILD"
+                   echo FINISHING BUILD
               else
-                   echo "Please specify cpu or gpu"
+                   echo Please specify cpu or gpu
                    )
 
               )
@@ -103,30 +105,30 @@ if  "%#" LSS 1  (
             else (
                   cmake  -DCPU_BLAS=true -DBLAS=TRUE ..
                   make && cd ..
-                  echo "FINISHING BUILD"
+                  echo FINISHING BUILD
                )
            )
 
 
       else (
-         if "%1" == "ptx" (
+         if %1 == ptx (
            rd /s /qptxbuild
            mkdir ptxbuild
            cd ptxbuild
            cmake -DPTX=TRUE ..
            make && cd ..
-           echo "FINISHING BUILD"
+           echo FINISHING BUILD
            move ptxbuild/ptx/cuda_compile_ptx_generated_all.cu.ptx all.ptx
            )
           )
      else (
-        if "%1" == "fatbin" (
+        if %1 == fatbin (
            rd /s /qfatbuild
            mkdir fatbuild
            cd fatbuild
            cmake -DFATBIN=TRUE ..
            make && cd ..
-           echo "FINISHING BUILD"
+           echo FINISHING BUILD
            move fatbuild/fatbin/cuda_compile_fatbin_generated_all.cu.fatbin all.fatbin
         )
 )
